requirements.txt
discord.py
bs4
aiohttp
asyncio
lxml
python-dotenv
